+++ pioreactor/automations/dosing/base.py
@@ -171,7 +171,9 @@ class DosingAutomationJob(AutomationJob):
 
     automation_name = "dosing_automation_base"  # is overwritten in subclasses
     job_name = "dosing_automation"
-    published_settings: dict[str, pt.PublishableSetting] = {}
+    published_settings: dict[str, pt.PublishableSetting] = {
+        "duration": {"datatype": "float", "settable": True}
+    }
 
     previous_normalized_od: Optional[float] = None
     previous_growth_rate: Optional[float] = None